git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@8421 01821946-851b-0410-b7f4-f1477c70e57a

+++ effects/lightning.cpp
@@ -804,7 +804,7 @@ while (nVerts-- > 0) {
 	if (zMin > v [Z])
 		zMin = v [Z];
 	}
-return 1.0f - zMin / ZRANGE;
+return 1.0f - zMin / (float) ZRANGE;
 }
 
 //------------------------------------------------------------------------------
